+++ extensions/shared/library/src/main/java/app/revanced/extension/shared/Utils.java
@@ -329,7 +329,7 @@ public class Utils {
             return (R) child;
         }
 
-        throw new IllegalArgumentException("View with resource name '" + str + "' not found");
+        throw new IllegalArgumentException("View with resource name not found: " + str);
     }
 
     /**