+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/videoplayer/CopyVideoUrlButton.java
@@ -7,7 +7,6 @@ import androidx.annotation.Nullable;
 import app.revanced.extension.shared.Logger;
 import app.revanced.extension.youtube.patches.CopyVideoUrlPatch;
 import app.revanced.extension.youtube.settings.Settings;
-import app.revanced.extension.youtube.shared.PlayerType;
 
 @SuppressWarnings("unused")
 public class CopyVideoUrlButton {
@@ -23,6 +22,7 @@ public class CopyVideoUrlButton {
                     controlsView,
                     "revanced_copy_video_url_button",
                     "revanced_copy_video_url_button_placeholder",
+                    null,
                     Settings.COPY_VIDEO_URL::get,
                     view -> CopyVideoUrlPatch.copyUrl(false),
                     view -> {