+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/videoplayer/ExternalDownloadButton.java
@@ -23,6 +23,7 @@ public class ExternalDownloadButton {
                     controlsView,
                     "revanced_external_download_button",
                     "revanced_external_download_button_placeholder",
+                    null,
                     Settings.EXTERNAL_DOWNLOADER::get,
                     ExternalDownloadButton::onDownloadClick,
                     null