+++ patches/src/main/kotlin/app/revanced/patches/instagram/misc/signature/SignatureCheckPatch.kt
@@ -8,7 +8,7 @@ val signatureCheckPatch = bytecodePatch(
     name = "Disable signature check",
     description = "Disables the signature check that causes the app to crash on startup."
 ) {
-    compatibleWith("com.instagram.android"("378.0.0.52.68"))
+    compatibleWith("com.instagram.android")
 
     execute {
         isValidSignatureMethodFingerprint