+++ patches/src/main/kotlin/app/revanced/patches/twitch/misc/settings/SettingsPatch.kt
@@ -48,7 +48,7 @@ val settingsPatch = bytecodePatch(
         settingsPatch(preferences = preferences),
     )
 
-    compatibleWith("tv.twitch.android.app")
+    compatibleWith("tv.twitch.android.app"("16.9.1"))
 
     execute {
         addResources("twitch", "misc.settings.settingsPatch")