+++ patches/src/main/kotlin/app/revanced/patches/youtube/video/hdr/DisableHdrPatch.kt
@@ -63,7 +63,7 @@ val disableHdrPatch = bytecodePatch(
                         return v0
                         :useHdr
                         nop 
-                     """
+                    """
                 )
             }
         }