+++ patches/src/main/kotlin/app/revanced/patches/youtube/video/quality/Fingerprints.kt
@@ -5,34 +5,6 @@ import app.revanced.util.literal
 import com.android.tools.smali.dexlib2.AccessFlags
 import com.android.tools.smali.dexlib2.Opcode
 
-internal const val YOUTUBE_VIDEO_QUALITY_CLASS_TYPE = "Lcom/google/android/libraries/youtube/innertube/model/media/VideoQuality;"
-
-internal val videoQualityFingerprint = fingerprint {
-    accessFlags(AccessFlags.PUBLIC, AccessFlags.CONSTRUCTOR)
-    parameters(
-        "I", // Resolution.
-        "Ljava/lang/String;", // Human readable resolution: "480p", "1080p Premium", etc
-        "Z",
-        "L"
-    )
-    custom { _, classDef ->
-        classDef.type == YOUTUBE_VIDEO_QUALITY_CLASS_TYPE
-    }
-}
-
-/**
- * Matches with the class found in [videoQualitySetterFingerprint].
- */
-internal val setVideoQualityFingerprint = fingerprint {
-    returns("V")
-    parameters("L")
-    opcodes(
-        Opcode.IGET_OBJECT,
-        Opcode.IPUT_OBJECT,
-        Opcode.IGET_OBJECT,
-    )
-}
-
 internal val videoQualityItemOnClickParentFingerprint = fingerprint {
     returns("V")
     strings("VIDEO_QUALITIES_MENU_BOTTOM_SHEET_FRAGMENT")
@@ -54,19 +26,6 @@ internal val videoQualityItemOnClickFingerprint = fingerprint {
     }
 }
 
-internal val videoQualitySetterFingerprint = fingerprint {
-    accessFlags(AccessFlags.PUBLIC, AccessFlags.FINAL)
-    returns("V")
-    parameters("[L", "I", "Z")
-    opcodes(
-        Opcode.IF_EQZ,
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.MOVE_RESULT_OBJECT,
-        Opcode.INVOKE_VIRTUAL,
-        Opcode.IPUT_BOOLEAN,
-    )
-    strings("menu_item_video_quality")
-}
 
 internal val videoQualityMenuOptionsFingerprint = fingerprint {
     accessFlags(AccessFlags.STATIC)